requirements.txt
streamlit
requests
pandas
matplotlib
groq
python-dotenv
feedparser